+++ src/main/java/ladysnake/dissolution/client/gui/GuiIncorporealOverlay.java
@@ -68,7 +68,7 @@ public class GuiIncorporealOverlay extends GuiIngameForge {
                     }
                 }
                 // We need to set the render view entity back to a player as renderAir and renderHotbar require it
-                // We shouldn't, though, if the entity has its own shader. This does resault in parts of the hotbar disappearing,
+                // We shouldn't, though, if the entity has its own shader. This does result in parts of the hotbar disappearing,
                 // but we'll fix that later...
                 boolean badMob;
                 boolean isCreeper = possessed instanceof EntityCreeper;