+++ src/main/java/ladysnake/dissolution/common/compat/MTSCompat.java
@@ -1,0 +1,11 @@
+package ladysnake.dissolution.common.compat;
+
+/**
+ * Adds compatibility with Immersive Vehicles (formerly MTS).
+ * @author sschr15
+ */
+public class /* enum */ MTSCompat {
+    // umm
+    // i dont know what im doing
+    // come back to me later
+}